fix(backends): allocate Qwen recurrent state at float32 even under bfloat16 (ASTRA-057)
The GDN kernels compute and return the recurrent state in float32 and DynamicCache preserves it
there even when weights/conv/KV are bfloat16. init_state allocated the recurrent buffer at the
model dtype, so a load(dtype=bfloat16) quantized it every update — first-chunk logits matched but
the second/third diverged (0.44 / 0.19), which a one-prefill parity test misses. The recurrent
memory is now allocated at float32 separately from conv/KV (which follow the model dtype),
restoring native parity. Verified on the real checkpoint: bf16 load keeps recurrent float32 across
three chunks, conv bf16. Regression test covers the multi-chunk bf16 dtype/continuation.

+++ plastic/backends/qwen.py
@@ -186,8 +186,12 @@ class QwenBackend:
                     dtype=self.dtype,
                 )
                 memory = torch.zeros(
+                    # the GDN kernels compute and return the recurrent state in float32 and
+                    # DynamicCache preserves it there even when weights/conv/KV are bfloat16;
+                    # allocating this at the model dtype would quantize it every update and diverge
+                    # after the first chunk (ASTRA-057). Conv/KV stay at the model dtype.
                     1, c.linear_num_value_heads, c.linear_key_head_dim, c.linear_value_head_dim,
-                    device=self.device, dtype=self.dtype,
+                    device=self.device, dtype=torch.float32,
                 )
                 layer.lazy_initialization(conv_states=conv, recurrent_states=memory)
                 layer.has_previous_state[0] = True
